+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * Added warning for setting cookies with empty names
 * Added color differentiation between error messages and warnings
 * Added detection of non-lowercase function names
+* Added the ```--relative-paths (-r)``` option to output file paths relative to checked directories
 
 ## Version 1.0.2
 * Fixed notice in ```PHP4ConstructorVisitor``` caused by anonymous classes